Revert receiver refactor
