+++ src/types/index.ts
@@ -5,3 +5,4 @@ export * from './command';
 export * from './events';
 export * from './options';
 export * from './view';
+export * from './receiver';
